Only replace feature localizations at boundaries
Ensure we do not replace a feature with a localized value if there are
letters or underscores (valid identifier characters) before/after it.
Because the \b regex considers underscores to be a part of a word we
cannot use it as our boundary and must check the character before or
after, if any, to make sure the identifier is not larger.