Replace organization branch filter/alter function with per-org config
Instead of allowing any function to be defined in the configuration, use
a mapping where the filter provides an expression to match and the alter
provides a similar expression to remove from the branch string.

This removes a potential vulnerability according to SonarQube (although
the source is trusted and Function scope is not the same as eval) and
makes the configuration cleaner by not mentioning other organizations in
the function to evaluate.